+++ NEWS.md
@@ -1,4 +1,4 @@
-# metR (development version)
+# metR 0.12.0
 
 ## New Features